Add files via upload
The documentation describes the creation of an employee manager application using Java Spring technology. The application uses the Spring Boot framework and the Java Spring MVC architecture to implement the server logic, Angular to implement the user interface on the front end, and a SQL database to store employee information. In this application, we can add employees, change their information, as well as delete users. There is also a search engine that makes it easier to find users on the platform. This application brings a number of advantages and efficiency in employee management. Here are some reasons why this project is good and effective for the organization:

Centralized data storage: The application uses a SQL database to store employee information. This enables centralized data storage, which facilitates access and management of employee information. All relevant data is available in one place, which facilitates decision-making and provides insight into all employees in the organization.

Process automation: Through the implementation of the Java Spring MVC architecture and the Spring Boot framework, the application enables the automation of processes related to employee management. This includes adding, editing and deleting employees, searching and filtering data, generating reports and other functionalities. Automating these processes reduces the need for manual work and increases the efficiency of business processes.

User experience: Angular is used to develop the front-end part of the application, resulting in a modern and intuitive user interface. Users will be able to easily navigate through the application, enter and update employee information, search for information and perform other actions. This improved user experience contributes to employee productivity and efficiency.

Scalability and adaptability: Using Java Spring MVVM architecture and Spring Boot framework enables easy scalability and adaptability of the application. If the organization grows and the number of employees increases, the application can be easily expanded and adapted to the needs. Also, adding new functionalities or changing existing ones is facilitated through a modular and flexible architecture.